requirements.txt
# Requirements for the project
# Use pip install -r requirements.txt to install all dependencies
obsws-python
colorama
pyyaml